@@ -3,27 +3,31 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emoji-mart/data": "^1.0.3",
+    "@emoji-mart/react": "^1.0.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
     "@fortawesome/free-brands-svg-icons": "^5.15.3",
     "@fortawesome/free-regular-svg-icons": "^5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
-    "@reduxjs/toolkit": "^1.6.0",
+    "@reduxjs/toolkit": "^1.8.3",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
+    "axios": "^0.27.2",
     "axios-mock-adapter": "^1.19.0",
-    "emoji-mart": "^3.0.1",
+    "emoji-mart": "^5.2.0",
     "fetch-mock": "^9.11.0",
     "firebase": "^8.7.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-contenteditable": "^3.3.5",
-    "react-debounce-input": "^3.2.3",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
+    "react-debounce-input": "^3.3.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-script": "^2.0.5",
     "react-scripts": "4.0.1",
+    "react-test-render": "^1.1.2",
     "react-test-renderer": "^17.0.2",
     "redux-mock-store": "^1.5.4"
   },
