@@ -3,9 +3,7 @@ import './FaqPage.css';
 import {
   Link,
   Route,
-  BrowserRouter as Router,
-  Switch,
-  useRouteMatch,
+  Routes,
 } from "react-router-dom";
 
 import AboutFaq from './AboutFaq';
@@ -16,45 +14,37 @@ import SocialFaq from './SocialFaq';
 
 export default function FaqPage() {
 
-  const match = useRouteMatch();
-
   return (
     <main id="faq-page">
       <div id="faq-container">
-        <Router>
             <div id="faq-questions">
               <ul>
-                <li><Link to={`${match.url}/About`}>What is Fomo App about?</Link></li>
+                <li><Link to={`About`}>What is Fomo App about?</Link></li>
                 <li>
                   <p>Features</p>
                   <ul>
-                    <li><Link to={`${match.url}/Portfolio`}>Portfolio Management</Link></li>
-                    <li><Link to={`${match.url}/LeaderBoards`}>Leader Boards</Link></li>
-                    <li><Link to={`${match.url}/Social`}>Social Features</Link></li>
+                    <li><Link to={`Portfolio`}>Portfolio Management</Link></li>
+                    <li><Link to={`LeaderBoards`}>Leader Boards</Link></li>
+                    <li><Link to={`Social`}>Social Features</Link></li>
                   </ul>
                 </li>
               </ul>
           </div>
           <div id="faq-answers">
-            <Switch>
-              <Route path={`${match.path}/About`}>
-                <AboutFaq />
+            <Routes>
+              <Route path={`About`} element={<AboutFaq/>}>
+                
               </Route>
-              <Route path={`${match.path}/Portfolio`}>
-                <PortfolioFaq />
+              <Route path={`Portfolio`} element={<PortfolioFaq/>}>
               </Route>
-              <Route path={`${match.path}/LeaderBoards`}>
-                <LeaderBoardFaq />
+              <Route path={`/LeaderBoards`} element={<LeaderBoardFaq/>}>
               </Route>
-              <Route path={`${match.path}/Social`}>
-                <SocialFaq />
+              <Route path={`/Social`} element={<SocialFaq/>}>
               </Route>
-              <Route path={`${match.path}/`}>
-                <AboutFaq />
+              <Route path={`/*`} element={<AboutFaq/>}>
               </Route>
-            </Switch>
+            </Routes>
           </div>
-        </Router>
       </div>
     </main>
   );
