@@ -2,29 +2,29 @@ import './index.css';
 
 import * as serviceWorker from './serviceWorker';
 
-import { Redirect, Route, BrowserRouter as Router, Switch } from 'react-router-dom';
+import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
 
 import App from './App';
 import { Provider } from 'react-redux';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from "react-dom/client";
 import RegistrationPage from './features/registration/RegistrationPage'
 import { store } from './app/Store';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
     <Provider store={store}>
       <Router>
-        <Switch>
-          <Route exact path="/" component={App}/>
-          <Route exact path="/portfolio/:urlUserId" component={App}/>
-          <Route exact path="/Register" component={RegistrationPage}/>
-          <Redirect to='/' />
-        </Switch>
+        <Routes>
+          <Route path="/" element={<App/>}/>
+          <Route path="portfolio/:urlUserId" element={<App/>}/>
+          <Route path="Register" element={<RegistrationPage/>}/>
+          <Route path="*" element={<App/>}/>
+        </Routes>
       </Router>
     </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>  
 );
 
 // If you want your app to work offline and load faster, you can change
